+++ src/test/java/com/znsio/teswiz/steps/JsonPlaceHolderSteps.java
@@ -1,0 +1,49 @@
+package com.znsio.teswiz.steps;
+
+import com.context.SessionContext;
+import com.context.TestExecutionContext;
+import com.znsio.teswiz.businessLayer.restUser.JsonPlaceHolderBL;
+import io.cucumber.java.en.And;
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+import kong.unirest.json.JSONObject;
+import org.apache.log4j.Logger;
+
+public class JsonPlaceHolderSteps {
+    private static final Logger LOGGER = Logger.getLogger(WeatherAPISteps.class.getName());
+    private final TestExecutionContext context;
+    private JSONObject jsonObject;
+    private int statusCode;
+
+    public JsonPlaceHolderSteps() {
+        context = SessionContext.getTestExecutionContext(Thread.currentThread().getId());
+        LOGGER.info("context: " + context.getTestName());
+    }
+    
+    @Given("I create a new post")
+    public void iCreateAPost() {
+        new JsonPlaceHolderBL().createPost();
+    }
+    @And("I modify title of the created post")
+    public void iUpdateTitleOfThePost() {
+        jsonObject = new JsonPlaceHolderBL().updatePost();
+    }
+
+    @Then("the title of the post should be updated")
+    public void verifyTitleUpdated() {
+        new JsonPlaceHolderBL().verifyPostUpdatedSuccessfully(jsonObject);
+
+    }
+
+    @When("I delete the modified post")
+    public void iDeleteThePost() {
+        statusCode= new JsonPlaceHolderBL().deletePost();
+    }
+
+
+    @Then("the post should be deleted successfully")
+    public void thePostShouldBeDeletedSuccessfully() {
+        new JsonPlaceHolderBL().verifyIfPostDeleted(statusCode);
+    }
+}
